drowsiness_model/requirements.txt
# Drowsiness Detection System — Dependencies
# ============================================
# Install: pip install -r requirements.txt
# Python 3.8+

# Core CV
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0   # for extra DNN modules

# Image handling (tkinter image display)
Pillow>=10.0.0

# Numerical
numpy>=1.24.0

# Face landmarks (OPTIONAL — but highly recommended for EAR accuracy)
# If dlib fails to install: pip install dlib --no-cache-dir
dlib>=19.24.0

# Notebook
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0            # for threshold analysis

# Testing
pytest>=7.4.0
